Seperated each function into its own Azure function and organized them in a directory
