+++ samples/PubSubSamples/PubSubDemo/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+    "PubSubDemo": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7060",
+      "launchBrowser": false
+    }
+  }
+}
