+++ samples/PubSubSamples/ReadThroughDemo/host.json
@@ -1,0 +1,11 @@
+{
+    "version": "2.0",
+    "logging": {
+      "applicationInsights": {
+        "samplingExcludedTypes": "Request",
+        "samplingSettings": {
+          "isEnabled": true
+        }
+      }
+    }
+}
