+++ samples/PubSubSamples/RedisFunctionsDemo/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+    "RedisFunctionsDemo": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7289",
+      "launchBrowser": false
+    }
+  }
+}
