+++ samples/PubSubSamples/WriteAroundDemo/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+    "WriteAroundDemo": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7080",
+      "launchBrowser": false
+    }
+  }
+}
