+++ samples/PubSubSamples/WriteBehindDemo/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+    "WriteBehindDemo": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7157",
+      "launchBrowser": false
+    }
+  }
+}
