+++ samples/PubSubSamples/WriteThroughDemo/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+    "WriteThroughDemo": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7149",
+      "launchBrowser": false
+    }
+  }
+}
